refactor - move TimeBarPlot to new class
